# src/main/java/com/servec/governor/api/UserApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.10).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.servec.governor.api;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.servec.governor.models.BadRequest;
import com.servec.governor.models.Login;
import com.servec.governor.models.User;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-12-22T04:22:26.530Z")

@Api(value = "user", description = "the user API")
public interface UserApi {

	@ApiOperation(value = "Logs user into the system", nickname = "loginUser", notes = "", response = User.class, tags = {
			"user", })
	@ApiResponses(value = { @ApiResponse(code = 200, message = "successful operation", response = User.class),
			@ApiResponse(code = 400, message = "Invalid username/password supplied") })
	@RequestMapping(value = "/user/login", produces = { "application/json" }, method = RequestMethod.POST)
	ResponseEntity<User> loginUser(@ApiParam(value = "", required = true) @Valid @RequestBody Login body);

	@ApiOperation(value = "Logs out current logged in user session", nickname = "logoutUser", notes = "", response = User.class, tags = {
			"user", })
	@ApiResponses(value = { @ApiResponse(code = 204, message = "Accepted", response = User.class),
			@ApiResponse(code = 400, message = "Bad Request", response = BadRequest.class),
			@ApiResponse(code = 404, message = "Not Found") })
	@RequestMapping(value = "/user/{username}/logout", produces = { "application/json" }, method = RequestMethod.PUT)
	ResponseEntity<User> logoutUser(
			@ApiParam(value = "The user name for login", required = true) @PathVariable("username") String username);

}