# src/main/java/com/servec/governor/api/VariableApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.10).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.servec.governor.api;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.servec.governor.models.BadRequest;
import com.servec.governor.models.Variable;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-12-22T04:22:26.530Z")

@Api(value = "variable", description = "the variable API")
public interface VariableApi {

	@ApiOperation(value = "Update existing Variable", nickname = "variableVariableIdPut", notes = "", response = Variable.class, tags = {
			"variable", })
	@ApiResponses(value = { @ApiResponse(code = 204, message = "No Content", response = Variable.class),
			@ApiResponse(code = 400, message = "Bad Request", response = BadRequest.class),
			@ApiResponse(code = 404, message = "Not Found") })
	@RequestMapping(value = "/variable/{variableId}", method = RequestMethod.PUT)
	ResponseEntity<Variable> variableVariableIdPut(
			@ApiParam(value = "", required = true) @PathVariable("variableId") Long variableId,
			@ApiParam(value = "Variable object", required = true) @Valid @RequestBody Variable body);

}